# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CartItem:
    properties:
      productID:
        type: string
      quantity:
        type: integer
    type: object
...
  dto.UpdateCartItemDTO:
    properties:
      quantity:
        type: integer
    type: object
...
  v1.failure:
    properties:
      error:
        $ref: '#/definitions/v1.failureInfo'
    type: object
  v1.failureInfo:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request body
        type: string
    type: object
  v1.success:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: paw1a@yandex.ru
    name: API Support
    url: https://t.me/paw1a
  description: This is simple api of e-commerce shop
  title: E-commerce API
  version: "1.0"
paths:
...
  /admins/carts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get all carts
      tags:
      - admin-carts
  /admins/carts/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: cart id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Delete cart
      tags:
      - admin-carts
    get:
      consumes:
      - application/json
      parameters:
      - description: cart id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get cart by id
      tags:
      - admin-carts
...
  /cart:
    delete:
      consumes:
      - application/json
      parameters:
      - description: cart id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Delete all cart items
      tags:
      - cart
    get:
      consumes:
      - application/json
      parameters:
      - description: cart id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Get cart items
      tags:
      - cart
    post:
      consumes:
      - application/json
      parameters:
      - description: cart item
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/domain.CartItem'
      - description: cart id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Add cart item
      tags:
      - cart
  /cart/{productID}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: productID
        required: true
        type: string
      - description: cart id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Delete cart item
      tags:
      - cart
    put:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: productID
        required: true
        type: string
      - description: cart item
        in: body
        name: cartItem
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartItemDTO'
      - description: cart id
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Update cart item
      tags:
      - cart
...
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"